backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
lxml==5.3.0
language-tool-python==2.7.1
python-docx==1.1.2
